Добавил картинок и разбил проект на три файла
