+++ lib/fnc/app.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+import 'package:lobanov_a_a_flutter_zadanie3/UI/homep.dart';
+
+class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      debugShowCheckedModeBanner: false,
+      title: 'Flutter Demo',
+      theme: ThemeData(
+        colorScheme: ColorScheme.fromSeed(seedColor: Colors.greenAccent),
+        useMaterial3: true,
+      ),
+      home: const MyHomePage(title: 'Пользователь: Лобанов Александр'),
+    );
+  }
+}
