+++ pubspec.yaml
@@ -59,9 +59,11 @@ flutter:
   uses-material-design: true
 
   # To add assets to your application, add an assets section, like this:
-  # assets:
-  #   - images/a_dot_burr.jpeg
-  #   - images/a_dot_ham.jpeg
+  assets:
+    - assets/images/bobtail.jpg
+    - assets/images/mainkun.jpg
+    - assets/images/Scotland.jpg
+    - assets/images/йода.jpg
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware
